- merges BaseGraphServiceClient and BaseClient
