+++ src/main/java/com/microsoft/graph/core/IBaseClient.java
@@ -25,6 +25,7 @@ package com.microsoft.graph.core;
 import javax.annotation.Nullable;
 import javax.annotation.Nonnull;
 
+import com.google.gson.JsonObject;
 import com.microsoft.graph.http.IHttpProvider;
 import com.microsoft.graph.logger.ILogger;
 import com.microsoft.graph.serializer.ISerializer;
@@ -76,4 +77,24 @@ public interface IBaseClient {
      * Validates this client
      */
     void validate();
+
+    /**
+     * Gets a builder to execute a custom request
+     *
+     * @return the custom request builder
+     * @param url the url to send the request to
+     * @param responseType the class to deserialize the response to
+     * @param <T> the type to deserialize the response to
+     */
+    @Nonnull
+    <T> CustomRequestBuilder<T> customRequest(@Nonnull final String url, @Nonnull final Class<T> responseType);
+
+    /**
+     * Gets a builder to execute a custom request with a generic JSONObject response
+     *
+     * @return the custom request builder
+     * @param url the url to send the request to
+     */
+    @Nonnull
+    CustomRequestBuilder<JsonObject> customRequest(@Nonnull final String url);
 }
